Bug 565581 - Fix issues reported by Sonar

+++ org.eclipse.nebula.widgets.nattable.core/src/org/eclipse/nebula/widgets/nattable/columnChooser/ColumnChooser.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012, 2019 Original authors and others.
+ * Copyright (c) 2012, 2020 Original authors and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -269,7 +269,7 @@ public class ColumnChooser {
                     ILayerCommand command = null;
                     if (!columnGroupMoved && !multipleColumnsMoved) {
                         int fromPosition = fromPositions.get(i).get(0).intValue();
-                        int toPosition = adjustToPosition(direction, toPositions.get(i).intValue());
+                        int toPosition = adjustToPosition(direction, toPositions.get(i));
                         command = new ColumnReorderCommand(ColumnChooser.this.columnHideShowLayer, fromPosition, toPosition);
                     } else if (columnGroupMoved && multipleColumnsMoved) {
                         command = new ReorderColumnsAndGroupsCommand(ColumnChooser.this.columnHideShowLayer, fromPositions.get(i),
