+++ org.eclipse.nebula.widgets.nattable.core/src/org/eclipse/nebula/widgets/nattable/painter/cell/VerticalTextPainter.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012, 2019 Original authors and others.
+ * Copyright (c) 2012, 2020 Original authors and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -272,7 +272,7 @@ public class VerticalTextPainter extends AbstractTextPainter {
 
                                 firstline = false;
                             } else {
-                                transform.translate(xOffset - previousXOffset, fontHeight + this.lineSpacing);
+                                transform.translate((float) xOffset - previousXOffset, (float) fontHeight + this.lineSpacing);
                             }
 
                             previousXOffset = xOffset;
@@ -300,7 +300,7 @@ public class VerticalTextPainter extends AbstractTextPainter {
 
                                 firstline = false;
                             } else {
-                                transform.translate(xOffset - previousXOffset, -fontHeight - this.lineSpacing);
+                                transform.translate((float) xOffset - previousXOffset, (float) -fontHeight - this.lineSpacing);
                             }
 
                             previousXOffset = xOffset;
