+++ org.eclipse.nebula.widgets.nattable.core/src/org/eclipse/nebula/widgets/nattable/painter/cell/decorator/PercentageBarDecorator.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012, 2015 Original authors and others.
+ * Copyright (c) 2012, 2020 Original authors and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -71,8 +71,8 @@ public class PercentageBarDecorator extends CellPainterWrapper {
                 Display.getCurrent(),
                 rectangle.x,
                 rectangle.y,
-                rectangle.x + rectangle.width,
-                rectangle.y + rectangle.height,
+                (float) rectangle.x + (float) rectangle.width,
+                (float) rectangle.y + (float) rectangle.height,
                 color1,
                 color2);
         gc.setBackgroundPattern(pattern);
